Move test for set-uuid-state into tests
